Revert "Binary compatible compact phrase table"
This reverts commit ad8fbc807e32ee7767dd72580bd71777c8714c01.

+++ moses/CompactPT/LexicalReorderingTableCompact.cpp
@@ -67,7 +67,7 @@ std::vector<float> LexicalReorderingTableCompact::GetScore(const Phrase& f,
       key = MakeKey(f,e,sub_c);
     }
     
-  uint64_t index = m_hash[key];
+  size_t index = m_hash[key];
   if(m_hash.GetSize() != index)
   {
     std::string scoresString;
